+++ src/main/java/domain/participant/Dealer.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.participant;
+
+import domain.card.Card;
 
 public class Dealer extends Participant {
