+++ src/main/java/domain/participant/Player.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import java.util.regex.Pattern;
