+++ src/main/java/domain/participant/Players.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.participant;
 
 import java.util.*;
